+++ src/models/Settings.php
@@ -27,4 +27,8 @@ class Settings extends Model
     public mixed $policies = [];
     
     public string $scriptTemplate = '';
+
+    public string $mapsApiKey = '';
+
+    public string $recaptchaSecretKey = '';
 }